third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/script-formatter-breakpoints-2.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script src="../debugger/resources/unformatted2.js"></script>
<script>
var test = function()
{
    Bindings.breakpointManager._storage._breakpoints = {};
    var panel = UI.panels.sources;
    var scriptFormatter;
    var formattedSourceFrame;

    InspectorTest.runDebuggerTestSuite([
        function testSetup(next)
        {
            InspectorTest.scriptFormatter().then(function(sf) {
                scriptFormatter = sf;
                next();
            });
        },

        function testBreakpointsSetAndRemoveInFormattedSource(next)
        {
            InspectorTest.showScriptSource("unformatted2.js", didShowScriptSource);

            function didShowScriptSource(frame)
            {
                InspectorTest.addSniffer(Sources.ScriptFormatterEditorAction.prototype, "_updateButton", uiSourceCodeScriptFormatted);
                scriptFormatter._toggleFormatScriptSource();
            }
                
            function uiSourceCodeScriptFormatted()
            {
                formattedSourceFrame = panel.visibleView;
                InspectorTest.setBreakpoint(formattedSourceFrame, 3, "", true);
                InspectorTest.waitBreakpointSidebarPane().then(evaluateF2);
            }

            function evaluateF2() {
                InspectorTest.waitUntilPaused(pausedInF2);
                InspectorTest.evaluateInPageWithTimeout("f2()");
            }

            function pausedInF2(callFrames)
            {
                InspectorTest.dumpBreakpointSidebarPane("while paused in pretty printed");
                InspectorTest.waitBreakpointSidebarPane()
                    .then(() => InspectorTest.dumpBreakpointSidebarPane("while paused in raw"))
                    .then(() => InspectorTest.resumeExecution(next));
                InspectorTest.removeBreakpoint(formattedSourceFrame, 3);
                scriptFormatter._discardFormattedUISourceCodeScript(panel.visibleView.uiSourceCode());
            }
        }
    ]);

}
</script>
</head>
<body onload="runTest()">
<p>Tests the script formatting is working fine with breakpoints.
</p>
</body>
</html>
