third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/script-formatter-breakpoints-2-expected.txt

Tests the script formatting is working fine with breakpoints.


Running: testSetup

Running: testBreakpointsSetAndRemoveInFormattedSource
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed

unformatted2.js:formatted:4    var b = 1;
Breakpoint sidebar pane while paused in raw

Script execution resumed.
